Moved packet tests to correct package
