+++ VM/src/test/java/uk/iatom/iAtomSys/common/api/TestLoadRequestPacket.java
@@ -1,10 +1,9 @@
-package uk.iatom.iAtomSys.common;
+package uk.iatom.iAtomSys.common.api;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import uk.iatom.iAtomSys.common.api.LoadRequestPacket;
 
 public class TestLoadRequestPacket {
